chore(mt-info-10): update uint to uint256 for consistency (#17)

+++ src/ArcadeStakingRewards.sol
@@ -665,7 +665,7 @@ contract ArcadeStakingRewards is IArcadeStakingRewards, ArcadeRewardsRecipient, 
         // This keeps the reward rate in the right range, preventing overflows due to
         // very high values of rewardRate in the earned and rewardsPerToken functions;
         // Reward + leftover must be less than 2^256 / 10^18 to avoid overflow.
-        uint balance = rewardsToken.balanceOf(address(this));
+        uint256 balance = rewardsToken.balanceOf(address(this));
 
         if (rewardRate > balance) revert ASR_RewardTooHigh();
